+++ name_criterion.py
@@ -1,5 +1,8 @@
 def name_criterion(name):
-    if len(name) > 20:
+    if not name.strip():  # Check if name is empty or only whitespace!! ;-))
+        print("User name cannot be enpty")
+        return None
+    elif len(name) > 20:
         print("User name should be within 20 characters")
-        return name
+        return None
     return name
